requirements.txt
cloudpickle==1.3.0
gym==0.17.2
gym-retro==0.8.0
numpy==1.18.5
pyglet==1.5.6
scipy==1.4.1
